+++ feature/characters/ui/src/iosMain/kotlin/cz/cernilovsky/kmp/rickandmorty/characters/RickAndMortySdkWidget.ios.kt
@@ -1,7 +1,6 @@
 package cz.cernilovsky.kmp.rickandmorty.characters
 
 import cz.cernilovsky.kmp.rickandmorty.characters.di.charactersUiModule
-import cz.cernilovsky.kmp.rickandmorty.core.annotation.InternalRickAndMortyApi
 import cz.cernilovsky.kmp.rickandmorty.runtime.RickAndMortySdk
 import cz.cernilovsky.kmp.rickandmorty.runtime.RickAndMortySdkConfig
 import cz.cernilovsky.kmp.rickandmorty.runtime.SdkMode
@@ -14,7 +13,6 @@ import org.koin.core.module.Module
  * Automatically registers [charactersUiModule]. Prefer this over plain [initialize] when hosting
  * screens from this artifact.
  */
-@OptIn(InternalRickAndMortyApi::class)
 public fun RickAndMortySdk.initializeWidget(
     config: RickAndMortySdkConfig = RickAndMortySdkConfig.builder().mode(SdkMode.Widget).build(),
     extraModules: List<Module> = emptyList(),