+++ runtime/src/commonMain/kotlin/cz/cernilovsky/kmp/rickandmorty/runtime/RickAndMortyFeatureFlags.kt
@@ -1,8 +1,6 @@
 package cz.cernilovsky.kmp.rickandmorty.runtime
 
-import cz.cernilovsky.kmp.rickandmorty.core.featureflags.domain.FeatureFlag
-
 /** Stable flag keys for [RickAndMortySdkConfig.Builder.overrideFeatureFlag]. */
 public object RickAndMortyFeatureFlags {
-    public val CHARACTER_DETAIL_AUTO_REFRESH: String = FeatureFlag.CharacterDetailAutoRefresh.key
+    public const val CHARACTER_DETAIL_AUTO_REFRESH: String = "character_detail_auto_refresh"
 }